--- a/src/gadgets/encode_image.rs
+++ b/src/gadgets/encode_image.rs
@@ -9,7 +9,10 @@ use ndarray::{Array2, Array3};
 
 use crate::gadgets::greater_than::GreaterThanWitnessResult;
 
-use super::greater_than::{GreaterThanChip, GreaterThanChipConfig, GreaterThanInstructions};
+use super::{
+    greater_than::{GreaterThanChip, GreaterThanChipConfig, GreaterThanInstructions},
+    range_check::RangeCheckConfig,
+};
 
 pub trait EncodeImageInstructions<F: PrimeFieldBits> {
     /// Maps an image to a bit string.
@@ -21,9 +24,9 @@ pub trait EncodeImageInstructions<F: PrimeFieldBits> {
 }
 
 #[derive(Clone, Debug)]
-pub struct EncodeImageChipConfig {
+pub struct EncodeImageChipConfig<F: PrimeFieldBits> {
     advice_column: Column<Advice>,
-    greater_than_chip_config: GreaterThanChipConfig,
+    greater_than_chip_config: GreaterThanChipConfig<F>,
 }
 
 /// Encodes an image into a bit string, as follows:
@@ -34,12 +37,15 @@ pub struct EncodeImageChipConfig {
 /// - Intensities belonging to the same pixel are constrained to be equal.
 pub struct EncodeImageChip<F: PrimeFieldBits> {
     greater_than_chip: GreaterThanChip<F>,
-    config: EncodeImageChipConfig,
+    config: EncodeImageChipConfig<F>,
     binarization_thresholds: Array3<u16>,
 }
 
 impl<F: PrimeFieldBits> EncodeImageChip<F> {
-    pub fn construct(config: EncodeImageChipConfig, binarization_thresholds: Array3<u16>) -> Self {
+    pub fn construct(
+        config: EncodeImageChipConfig<F>,
+        binarization_thresholds: Array3<u16>,
+    ) -> Self {
         let greater_than_chip = GreaterThanChip::construct(config.greater_than_chip_config.clone());
         Self {
             greater_than_chip,
@@ -55,9 +61,10 @@ impl<F: PrimeFieldBits> EncodeImageChip<F> {
         diff: Column<Advice>,
         is_gt: Column<Advice>,
         byte_column: TableColumn,
-    ) -> EncodeImageChipConfig {
+        range_check_config: RangeCheckConfig<F>,
+    ) -> EncodeImageChipConfig<F> {
         let greater_than_chip_config =
-            GreaterThanChip::configure(meta, x, y, diff, is_gt, byte_column);
+            GreaterThanChip::configure(meta, x, y, diff, is_gt, byte_column, range_check_config);
         EncodeImageChipConfig {
             advice_column: is_gt,
             greater_than_chip_config,
@@ -110,7 +117,7 @@ impl<F: PrimeFieldBits> EncodeImageInstructions<F> for EncodeImageChip<F> {
                                 // add a copy constraint for the other thresholds.
                                 let GreaterThanWitnessResult { x_cell, gt_cell } =
                                     self.greater_than_chip.greater_than_witness(
-                                        &mut layouter,
+                                        layouter.namespace(|| format!("gt[{}, {}]", i, j)),
                                         F::from(image[(i, j)] as u64),
                                         t,
                                     )?;
@@ -120,7 +127,7 @@ impl<F: PrimeFieldBits> EncodeImageInstructions<F> for EncodeImageChip<F> {
                             Some(first_cell) => {
                                 // For the other cells, we want to add a copy constraint to the first cell.
                                 self.greater_than_chip.greater_than_copy(
-                                    &mut layouter,
+                                    layouter.namespace(|| format!("gt[{}, {}]", i, j)),
                                     first_cell,
                                     t,
                                 )?