… examples/evm-verifier.rs
… models/readme.md
… src/gadgets/wnn.rs
… src/lib.rs
… tests/using_zero_g_as_a_library.rs